requirements.txt
Django==4.2.17
djangorestframework==3.15.1
django-cors-headers
Pillow